StorageProvider Fix: move isKeyspaceWithLocalStrategy to SchemaManager
