+++ src/java/org/apache/cassandra/schema/SchemaConstants.java
@@ -27,7 +27,6 @@ import java.util.regex.Pattern;
 import com.google.common.collect.ImmutableSet;
 
 import org.apache.cassandra.db.Digest;
-import org.apache.cassandra.locator.LocalStrategy;
 
 public final class SchemaConstants
 {
@@ -84,29 +83,6 @@ public final class SchemaConstants
     public static boolean isLocalSystemKeyspace(String keyspaceName)
     {
         return LOCAL_SYSTEM_KEYSPACE_NAMES.contains(keyspaceName.toLowerCase());
-    }
-
-    /**
-     * @return whether or not the keyspace is a really system one (w/ LocalStrategy, unmodifiable, hardcoded)
-     *        or it's having {@link LocalStrategy}
-     */
-    public static boolean isKeyspaceWithLocalStrategy(String keyspaceName)
-    {
-        return isLocalSystemKeyspace(keyspaceName) ||
-               (SchemaManager.instance.getKeyspaceMetadata(keyspaceName) != null
-                && SchemaManager.instance.getKeyspaceMetadata(keyspaceName).params.replication.klass.equals(LocalStrategy.class));
-    }
-
-    /**
-     * Equivalent to {@link #isKeyspaceWithLocalStrategy(String)} but uses the provided keyspace metadata instead
-     * of getting the metadata from the schema manager
-     *
-     * @param keyspace the keyspace metadata to check
-     * @return if the provided keyspace uses local replication strategy
-     */
-    public static boolean isKeyspaceWithLocalStrategy(KeyspaceMetadata keyspace)
-    {
-        return isLocalSystemKeyspace(keyspace.name) || keyspace.params.replication.klass.equals(LocalStrategy.class);
     }
 
     /**
