+++ src/java/org/apache/cassandra/schema/SchemaManager.java
@@ -800,4 +800,26 @@ public final class SchemaManager implements SchemaProvider
                : Collections.emptyMap();
     }
 
+    /**
+     * @return whether or not the keyspace is a really system one (w/ LocalStrategy, unmodifiable, hardcoded)
+     *        or it's having {@link LocalStrategy}
+     */
+    public static boolean isKeyspaceWithLocalStrategy(String keyspaceName)
+    {
+        return SchemaConstants.isLocalSystemKeyspace(keyspaceName) ||
+               (instance.getKeyspaceMetadata(keyspaceName) != null
+                && instance.getKeyspaceMetadata(keyspaceName).params.replication.klass.equals(LocalStrategy.class));
+    }
+
+    /**
+     * Equivalent to {@link #isKeyspaceWithLocalStrategy(String)} but uses the provided keyspace metadata instead
+     * of getting the metadata from the schema manager
+     *
+     * @param keyspace the keyspace metadata to check
+     * @return if the provided keyspace uses local replication strategy
+     */
+    public static boolean isKeyspaceWithLocalStrategy(KeyspaceMetadata keyspace)
+    {
+        return SchemaConstants.isLocalSystemKeyspace(keyspace.name) || keyspace.params.replication.klass.equals(LocalStrategy.class);
+    }
 }
